\name{myapp-package}
\alias{myapp-package}
\alias{myapp}
\docType{package}
\title{
\packageTitle{myapp}
}
\description{
\packageDescription{myapp}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{myapp}
\packageIndices{myapp}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{myapp}

Maintainer: \packageMaintainer{myapp}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
# simple examples of the most important functions
}
