% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myFun.R
\name{myFun}
\alias{myFun}
\title{Multiplica 2 numeros double}
\usage{
myFun(x, y)
}
\arguments{
\item{x}{Un valor double}

\item{y}{Un valor double}
}
\value{
Un valor double
}
\description{
Multiplica 2 numeros double
}
\examples{
myFun(13,7)
}
